requirements.txt
asgiref==3.11.1
joblib
numpy
pandas
catboost
scikit-learn==1.6.1
qdrant-client
python-dotenv
Django==5.1.15
django-cors-headers==4.9.0
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
psycopg2-binary==2.9.11
PyJWT==2.12.1
sqlparse==0.5.5
tzdata==2026.1
